fit entity

+++ AISEP/src/AISEP.BLL/DTOs/Responses/NonPremiumProjectResponse.cs
@@ -15,7 +15,7 @@ namespace AISEP.BLL.DTOs.Responses
         public ProjectScorecardDto? ProjectScorecard { get; set; }
         public string? ProjectImageUrl { get; set; }
         public string Status { get; set; } = string.Empty;
-        public int? StartupPotentialScore { get; set; }
+        public decimal? StartupPotentialScore { get; set; }
         public DateTime CreatedAt { get; set; }
         public int FollowerCount { get; set; }
         public bool IsFollowedByCurrentUser { get; set; }
